requirements.txt
praw
python-dotenv
langchain
langchain-google-genai